Line number lookup
Add API to out which file/line the current instruction is for. I don't
think we need to justify why that is useful

Implement the line program machine according to the dwarf 4 spec until
we can process our debugee program successfully. Note that libdwarf does
this for us, however it is more fun and interesting to do it ourselves,
so we do

Implementation details:
* Current implementation is pretty naive, it reruns the whole line
  program for a compilation unit every time a source location is looked
  up. This is easily optimizable in the future (save machine state for
  each fn or something), but for now is good enough
* Track compilation units to figure out which line program to run
* There are some fixme's scattered around. I'm fine to leave them in to
  deal with them later. This seems to work well enough for now